requirements.txt
flask==3.0.3
dash==2.17.1
plotly==5.22.0
pandas==2.2.2
mysql-connector-python==8.4.0
statsmodels==0.14.2
